# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Viva Challenge Code
host: localhost:10010
basePath: /v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /properties:
     x-swagger-router-controller: property
     post:
       description: create a specific property in Spotippos
       operationId: postProperty
       parameters:
        - name: property
          in: body
          required: true
          schema:
            $ref: "#/definitions/Property"
       responses:
         "200":
           description: Success
           schema:
             $ref: "#/definitions/GeneralResponse"
         default:
           description: Error
           schema:
             $ref: "#/definitions/ErrorResponse"
     get:
       description: search properties in Spotippos
       operationId: getProperties
       parameters:
        - name: ax
          type: integer
          in: query
          required: true
          minimum: 0
          maximum: 1400
        - name: ay
          type: integer
          in: query
          required: true
          minimum: 0
          maximum: 1000
        - name: bx
          type: integer
          in: query
          required: true
          minimum: 0
          maximum: 1400
        - name: by
          type: integer
          in: query
          required: true
          minimum: 0
          maximum: 1000
       responses:
         "200":
           description: Success
           schema:
             $ref: "#/definitions/GetPropertiesResponse"
         "404":
             description: NotFound
             schema:
               $ref: "#/definitions/NotFoundResponse"
         default:
           description: Error
           schema:
             $ref: "#/definitions/ErrorResponse"
  /properties/{id}:
     x-swagger-router-controller: property
     get:
       description: get a specific property in Spotippos
       operationId: getProperty
       parameters:
        - name: id
          type: string
          in: path
          required: true
       responses:
         "200":
           description: Success
           schema:
             $ref: "#/definitions/GetPropertyResponse"
         "404":
             description: NotFound
             schema:
               $ref: "#/definitions/NotFoundResponse"
         default:
           description: Error
           schema:
             $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  GetPropertiesResponse:
    required:
      - foundProperties
      - properties
    properties:
      foundProperties:
        type: integer
      properties:
        type: array
        items:
          $ref: "#/definitions/Property"
  GetPropertyResponse:
    required:
      - _id
      - title
      - price
      - x
      - y
      - beds
      - baths
      - provinces
      - squareMeters
    properties:
      _id:
        type: string
      title:
        type: string
      price:
        type: number
      x:
        type: number
      y:
        type: number
      beds:
        type: number
      baths:
        type: number
      provinces:
        type: array
        items:
          type: string
      squareMeters:
        type: number
  Property:
    required:
      - title
      - price
      - description
      - x
      - y
      - beds
      - baths
      - squareMeters
    properties:
      title:
        type: string
      price:
        type: integer
      x:
        type: integer
      y:
        type: integer
      beds:
        type: integer
      baths:
        type: integer
      squareMeters:
        type: integer
      description:
        type: string
  GeneralResponse:
    type: object
    required:
      - message
    properties:
      message:
        type: string
        description: a short comment
      entityId:
        type: string
        description: a short comment
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  NotFoundResponse:
    required:
      - message
    properties:
      message:
        type: string